+++ docs/CHANGE_LOG.md
@@ -1,7 +1,9 @@
 # JsonTree.js - Change Log:
 
 ## Version 0.2.0:
-- 
+
+#### **UI Improvements:**
+- The spacing used for the title bar buttons is now smaller when viewed in a mobile.
 
 <br>