+++ docs/CHANGE_LOG.md
@@ -6,6 +6,10 @@
 - Added a new binding option called "sortPropertyNamesInAlphabeticalOrder", which states if the sorted property names for an object should be in alphabetical order (defaults to true).
 - Added a new binding option called "showCommas", which states if commas should be shown at the end of each line (defaults to false).
 
+#### **Public Functions:**
+- Added new public function "render()", which will render a specific DOM element using the options you specify.
+- Added new public function "renderAll()", which will find all new DOM elements with the "data-jsontree-options" attribute and render them.
+
 #### **UI Improvements:**
 - The spacing used for the title bar buttons is now smaller when viewed in a mobile.